Make common_prefix_count unable to overflow

+++ fuel-merkle/src/common/path.rs
@@ -19,7 +19,7 @@ pub trait Path {
     /// Which child node to follow at the given index.
     fn get_instruction(&self, index: u32) -> Option<Side>;
 
-    fn common_path_length(&self, other: &[u8]) -> u32;
+    fn common_path_length(&self, other: &[u8]) -> u64;
 }
 
 impl<T> Path for T
@@ -30,7 +30,7 @@ where
         self.get_bit_at_index_from_msb(index).map(Into::into)
     }
 
-    fn common_path_length(&self, other: &[u8]) -> u32 {
+    fn common_path_length(&self, other: &[u8]) -> u64 {
         self.common_prefix_count(other)
     }
 }
